Add license and update basic readme information with more detail description of the component (#3)
As mentioned in TCB meeting, we use a open source license. But not yet decided which one exactly to use. EULA and GPL were mentioned as well.
Since there are not yet too much code and is under design phase, I use MIT at the moment and the decision need to be evaluated soon.

README is update to reflect more draft design decisions described in the RFCs.